feat: adicionada página de evento com a programação do fbis
